**New** N1. Merged replicates: now MuSERA automatically merges the output set of analyzed replicates into a single file; by merging the ERs and combing the p-values of overlapping ERs using Fisher's method. The output that is a single file per each session, is automatically exported to session folder root when the session is saved. N2. InstallSheild setup is added. N3. Option "-100xLog10 p-value" for p-value conversion is added.
**Updates**
U1. Batch execution reports whether the project is completed successfully or with a number of errors.
U2. Save dialog closes after save process is completed.
U3. Decimals of timers both in interactive and batch mode are reduced to 2 digits of millisecond.

**Bug Fixes**
B1. An issue with batch process abort on execution failure is resolved.
B2. Double-clicking on sets in details datagrid while the datagrid is empty (i.e., either an analysis is not executed, or no chromosome is selected for analyzed sample) were crashing MuSERA. This bug is resolved.
B3. A bug with changing samples color is resolved.
B4. A bug with batch mode triggered with a file containing wrong root structure is resolved.

**Minor Changes**
M1. A comment is added to sample at-Job file informing that the specified paths are "absolute".
M2. Date prefix in session folder name is changed.
M3. Priority descriptions in Batch execution are updated.
M4. Sign information are removed.
M4. Session execution description is updated.